A little descriptions of the game
This is a gomuku game, and this has two types of playing styles. First one is dual player and it can be play against each other. The second mode is involve with AI player, which means that even you are alone you can still play this game because there is an AI player in it. The game will be updated soon.